6369 | 7 files | M DumpSource2/schemas/client/C_DOTA_Ability_Brewmaster_CinderBrew.h,  M DumpSource2/schemas/server/CDOTA_Ability_Brewmaster_CinderBrew.h,  M Protobufs/dota_gcmessages_client_watch.proto,  M Protobufs/dota_gcmessages_common.proto,  M game/bin/built_from_cl.txt,  M game/dota/bin/linuxsteamrt64/libclient_strings.txt,  M game/dota/steam.inf
https://steamdb.info/patchnotes/17813524/

+++ DumpSource2/schemas/client/C_DOTA_Ability_Brewmaster_CinderBrew.h
@@ -2,6 +2,6 @@ class C_DOTA_Ability_Brewmaster_CinderBrew
 {
 	CUtlVector< CHandle< C_BaseEntity > > m_hUnitsHit;
 	float32 duration;
-	int32 barrel_impact_damage;
-	int32 barrel_width;
+	float32 barrel_impact_damage;
+	float32 barrel_width;
 };
